refactor(config): split irc helpers from channel config hotspot
